fix_loop_duplicates command: move logic to own file
